# prowl/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Prowl LLM Orchestration API",
    "description": "REST API for LLM-driven orchestration of the Prowl security crawler. Enables programmatic crawl control, discovery queries, mid-crawl adjustments, intervention handling, and attack surface reporting.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/endpoints": {
      "get": {
        "tags": [
          "v1",
          "discovery"
        ],
        "summary": "Query discovered endpoints",
        "description": "Returns endpoints with rich filtering. Supports filtering by tag, page_type, HTTP method, auth requirement, parameter presence, source module, and URL pattern.",
        "operationId": "list_endpoints_api_v1_endpoints_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "description": "Page size (max 500)",
              "default": 50,
              "title": "Limit"
            },
            "description": "Page size (max 500)"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Offset for pagination",
              "default": 0,
              "title": "Offset"
            },
            "description": "Offset for pagination"
          },
          {
            "name": "tag",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by tag (e.g. api, admin, debug)",
              "title": "Tag"
            },
            "description": "Filter by tag (e.g. api, admin, debug)"
          },
          {
            "name": "page_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by page type (login, error, api_json, admin, etc.)",
              "title": "Page Type"
            },
            "description": "Filter by page type (login, error, api_json, admin, etc.)"
          },
          {
            "name": "method",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by HTTP method (GET, POST, PUT, etc.)",
              "title": "Method"
            },
            "description": "Filter by HTTP method (GET, POST, PUT, etc.)"
          },
          {
            "name": "requires_auth",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by auth requirement",
              "title": "Requires Auth"
            },
            "description": "Filter by auth requirement"
          },
          {
            "name": "has_params",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by parameter presence",
              "title": "Has Params"
            },
            "description": "Filter by parameter presence"
          },
          {
            "name": "source_module",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by source module (s1_spider, s4_js, etc.)",
              "title": "Source Module"
            },
            "description": "Filter by source module (s1_spider, s4_js, etc.)"
          },
          {
            "name": "url_pattern",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Substring match on URL",
              "title": "Url Pattern"
            },
            "description": "Substring match on URL"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/endpoints/{fingerprint}": {
      "get": {
        "tags": [
          "v1",
          "discovery"
        ],
        "summary": "Get endpoint by fingerprint",
        "description": "Returns a single endpoint by its fingerprint hash.",
        "operationId": "get_endpoint_api_v1_endpoints__fingerprint__get",
        "parameters": [
          {
            "name": "fingerprint",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Fingerprint"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EndpointResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/input-vectors": {
      "get": {
        "tags": [
          "v1",
          "discovery"
        ],
        "summary": "Query input vectors",
        "description": "Returns input vectors (attack surface parameters) with filtering by risk indicator, location, reflection status, and endpoint URL.",
        "operationId": "list_input_vectors_api_v1_input_vectors_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "risk_indicator",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by risk tag (ssrf_candidate, sqli_candidate, cmdi_candidate, etc.)",
              "title": "Risk Indicator"
            },
            "description": "Filter by risk tag (ssrf_candidate, sqli_candidate, cmdi_candidate, etc.)"
          },
          {
            "name": "location",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by location (query, body, header, cookie, path)",
              "title": "Location"
            },
            "description": "Filter by location (query, body, header, cookie, path)"
          },
          {
            "name": "is_reflected",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by reflection status (XSS candidates)",
              "title": "Is Reflected"
            },
            "description": "Filter by reflection status (XSS candidates)"
          },
          {
            "name": "endpoint_url",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by endpoint URL substring",
              "title": "Endpoint Url"
            },
            "description": "Filter by endpoint URL substring"
          },
          {
            "name": "source_module",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by source module",
              "title": "Source Module"
            },
            "description": "Filter by source module"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tech-stack": {
      "get": {
        "tags": [
          "v1",
          "discovery"
        ],
        "summary": "Get detected technologies",
        "description": "Returns all detected technologies with confidence scores and evidence.",
        "operationId": "list_tech_stack_api_v1_tech_stack_get",
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by category (framework, server, cms, frontend, waf)",
              "title": "Category"
            },
            "description": "Filter by category (framework, server, cms, frontend, waf)"
          },
          {
            "name": "min_confidence",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "description": "Minimum confidence threshold",
              "default": 0.0,
              "title": "Min Confidence"
            },
            "description": "Minimum confidence threshold"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TechFingerprintResponse"
                  },
                  "title": "Response List Tech Stack Api V1 Tech Stack Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/secrets": {
      "get": {
        "tags": [
          "v1",
          "discovery"
        ],
        "summary": "Get discovered secrets",
        "description": "Returns all discovered secrets (API keys, tokens, passwords).",
        "operationId": "list_secrets_api_v1_secrets_get",
        "parameters": [
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by secret kind (api_key, token, etc.)",
              "title": "Kind"
            },
            "description": "Filter by secret kind (api_key, token, etc.)"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SecretResponse"
                  },
                  "title": "Response List Secrets Api V1 Secrets Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth-boundaries": {
      "get": {
        "tags": [
          "v1",
          "discovery"
        ],
        "summary": "Get auth boundaries",
        "description": "Returns authentication boundary transitions.",
        "operationId": "list_auth_boundaries_api_v1_auth_boundaries_get",
        "parameters": [
          {
            "name": "boundary_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by type (redirect_to_login, 403_forbidden, etc.)",
              "title": "Boundary Type"
            },
            "description": "Filter by type (redirect_to_login, 403_forbidden, etc.)"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuthBoundaryResponse"
                  },
                  "title": "Response List Auth Boundaries Api V1 Auth Boundaries Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/transactions/{txn_id}": {
      "get": {
        "tags": [
          "v1",
          "discovery"
        ],
        "summary": "Get transaction detail",
        "description": "Returns full transaction detail including headers and body sizes.",
        "operationId": "get_transaction_api_v1_transactions__txn_id__get",
        "parameters": [
          {
            "name": "txn_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Txn Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/transactions/{txn_id}/response-body": {
      "get": {
        "tags": [
          "v1",
          "discovery"
        ],
        "summary": "Get transaction response body",
        "description": "Returns the raw response body bytes for a specific transaction.",
        "operationId": "get_transaction_body_api_v1_transactions__txn_id__response_body_get",
        "parameters": [
          {
            "name": "txn_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Txn Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaginatedResponse": {
        "properties": {
          "items": {
            "items": {},
            "type": "array",
            "title": "Items",
            "description": "Page of results"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Total matching items"
          },
          "offset": {
            "type": "integer",
            "title": "Offset",
            "description": "Current offset"
          },
          "limit": {
            "type": "integer",
            "title": "Limit",
            "description": "Page size"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More",
            "description": "Whether more items exist beyond this page"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "offset",
          "limit",
          "has_more"
        ],
        "title": "PaginatedResponse",
        "description": "Cursor-based pagination wrapper."
      },
...
    }
  }
}